refactor: Have `TrackListeningSession` be the results of a function to create a shared scope
- This allows to use it in other places if we want to and affect the same data.
